Migrate to use PostgreSQL ... sqlx macros aren't great great

+++ db/migrations/0_steam.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS steam_games
-( id INTEGER NOT NULL
+( id BIGINT NOT NULL
 , name VARCHAR NOT NULL
-, upserted_at INTEGER NOT NULL
+, upserted_at BIGINT NOT NULL
 , PRIMARY KEY (id)
 );
 
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS steam_game_achievements
 , description VARCHAR
 , icon_locked_url VARCHAR NOT NULL
 , icon_unlocked_url VARCHAR NOT NULL
-, steam_game_id INTEGER NOT NULL
+, steam_game_id BIGINT NOT NULL
 , FOREIGN KEY (steam_game_id) REFERENCES steam_games(id) ON DELETE CASCADE
 , UNIQUE (steam_game_id, name)
 );
