+++ src/steam/api_steampowered_com/i_steam_apps/get_app_list/mod.rs
@@ -14,7 +14,7 @@ pub struct ResponseAppList {
 #[derive(Deserialize, Debug)]
 pub struct ResponseApp {
     #[serde(rename = "appid")]
-    pub app_id: u32,
+    pub app_id: i64,
 }
 
 pub struct Request {
